requirements.txt
numpy
pandas
scikit-learn
streamlit
requests
beautifulsoup4
nltk
joblib
python-dotenv
langchain
transformers